requirements.txt
alembic==1.13.2
annotated-types==0.7.0
APScheduler==3.10.4
blinker==1.8.2
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
Flask==2.1.1
Flask-Migrate==4.0.5
Flask-SQLAlchemy==2.5.1
greenlet==3.0.3
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
Mako==1.3.5
MarkupSafe==2.1.5
pydantic==2.5.2
pydantic_core==2.14.5
PyMySQL==1.1.0
pytz==2024.1
requests==2.31.0
six==1.16.0
SQLAlchemy==1.4.18
typing_extensions==4.12.2
tzlocal==5.2
urllib3==2.2.2
Werkzeug==2.3.8
PyJWT~=2.9.0